Team now contains number of wins per season and number of points per season, plus DateUtil.actualYear refactored to String and ranking menu shows last active season

+++ app/src/main/java/cz/prague/cvut/fit/steuejan/amtelapp/business/managers/MatchManager.kt
@@ -69,5 +69,5 @@ object MatchManager
     }
 
     fun getMatches(round: Int, group: String): Query
-            = MatchDAO().getMatches(round, group, DateUtil.actualYear)
+            = MatchDAO().getMatches(round, group, DateUtil.actualYear.toInt())
 }
